ci: 👷‍♀️ add pkgdown workflow (#16)
## Description

This PR adds the workflow for building the pkgdown website. The first
commit is a direct copy from `osdc` while the second commit updates this
workflow with how the [workflow
currently](https://github.com/r-lib/actions/blob/v2/examples/pkgdown.yaml)
is in the `r-lib` repo.

Let me know which one you prefer :) 

Related to #6 

This PR needs an in-depth review.

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>