refactor: :recycle: export get helper functions (#51)
# Description

I'm on a journey to get rid of the check errors, so it's easier to keep
track of whether newly introduced functions break things.

For now, since @lwjohnst86 don't remember whether the `get_*()` helper
functions should be internal or not, I'll export them to get rid of some
errors of them being used in examples.

Needs no review.

## Checklist

- [X] Ran `just run-all`