chore: :truck: move info from `CITATION.cff` into `DESCRIPTION` (#54)
# Description

EDIT: This started as a move of the CITATION file into the `inst/`
folder, but I found that we could just add the information to the
`DESCRIPTION` file (following what they do in the `dplyr` repo). See
comments below in the PR for elaboration.

This is the proper location for it according to the CMD checks and
https://r-pkgs.org/misc.html#sec-misc-inst.

Needs no review.

## Checklist

- [X] Ran `just run-all`