docs: :memo: update index page with overview, purpose, and why parquet sections (#63)
# Description

This PR updates and adds sections to the index page, including an
overview with link to the design vignette, a purpose section, and a
section on why it makes sense to convert to Parquet files.

This last section could potentially link to another page with more
elaboration, context, and examples of the advantages in the future (so
this index page doesn't get too much in-depth information). The section
was largely made with autocomplete/co-pilot and is just meant as an
initial version. I suspect you, @lwjohnst86, will have more/other
perspectives to add to that section.

Needs a thorough review.

## Checklist

- [X] Ran `just run-all`
- [X] If docs were added, Markdown is formatted

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>