docs: 📝 add functions section to design (#69)
# Description

This PR adds an overview of the functions we'll include in the package
(or an initial version thereof).
It also adds pseudocode for the conversion
function`convert_to_parquet()`
 
NOTE: I currently add all things design in the design vignette. It might
have become too long with the current PRs (#67 #68). Could be that we
should have a separate interface vignette?

Related to #7 

Needs a thorough review.

## Checklist

- [X] Ran `just run-all`
- [X] If docs were added, Markdown is formatted

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>