feat: ✨ add `read_register()` (#90)
# Description

This PR adds `read_register()` along with two internals for reading a
partitioned and non-partitioned Parquet.

This implementation is quite naive: If the given `path` is a directory,
it will try to load it as a partitioned Parquet with `open_dataset()`,
and if it's a file, it will try to load it as a Parquet file
(`read_parquet()`. Do we want something a bit more sophisticated than
this?

Closes #79 

Needs a thorough review.

## Checklist

- [X] Ran `just run-all`

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>